+++ src/server/src/projects/services/projects.service.ts
@@ -18,7 +18,8 @@ export class ProjectsService extends TypeOrmCrudService<Project> {
   constructor(@InjectRepository(Project) repo: Repository<Project>) {
     super(repo);
   }
-  save(project: Project): Promise<Project> {
+
+  save(project: Partial<Project>): Promise<Project> {
     // @ts-ignore
     return this.repo.save(project);
   }