+++ src/server/src/region-configs/controllers/region-configs.controller.ts
@@ -35,6 +35,9 @@ import * as _ from "lodash";
         persist: ["regionConfig"],
         eager: true
       }
+    },
+    filter: {
+      archived: false
     }
   },
   routes: {